fix(api): For issue/event search, remap column names that conflict with tags
We have a few cases where people have existing searches that search on `platform`, `type`,
`version`, etc. These are valid columns in snuba, but previously we treated them as tags when
searching on them. To not break existing searches, I'm remapping these columns to `<column>.name` in
our searches.

For reference, a list of these types of fields is available at
https://github.com/getsentry/snuba/blob/fcc0bc6be085afb0879904e99dd007409dea4e58/snuba/clickhouse.py#L441